USBManager: Prevent VC++ warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10944 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/USBManager/USBManager.cpp
@@ -548,7 +548,7 @@ int MDUSBDevice::Read(unsigned char* buf, unsigned long bufLen, unsigned long& c
          statusContinue = false;
       else if (charsReceived > 0)
       {
-         if (charsReceived > bufLen - charsRead) {
+         if (charsReceived > (int)(bufLen - charsRead)) {
             memcpy((unsigned char*) buf + charsRead, internalBuf, bufLen - charsRead);
             overflowBufferLength_ = charsReceived - bufLen + charsRead;
             // we have extra chars, save in persistent buffer
